+++ modern-job-board.php
@@ -3,7 +3,7 @@
  * Plugin Name: Modern Job Board
  * Plugin URI: https://github.com/MartinOrton/modern-job-board
  * Description: A lightweight job board plugin for WordPress.
- * Version: 1.8.3
+ * Version: 1.8.4
  * Author: Martin Orton
  * Author URI: https://www.martinorton.com
  * Text Domain: modern-job-board
@@ -14,7 +14,7 @@ if (!defined('ABSPATH')) {
 }
 
 // Define plugin constants.
-define('MJB_VERSION', '1.8.3');
+define('MJB_VERSION', '1.8.4');
 define('MJB_PATH', plugin_dir_path(__FILE__));
 define('MJB_URL', plugin_dir_url(__FILE__));
 
@@ -272,7 +272,8 @@ class Modern_Job_Board
             return;
         }
 
-        wp_enqueue_style('mjb-style', MJB_URL . 'assets/css/mjb-style.css', array(), MJB_VERSION);
+        wp_enqueue_style('mjb-shared', MJB_URL . 'assets/css/mjb-shared.css', array(), MJB_VERSION);
+        wp_enqueue_style('mjb-style', MJB_URL . 'assets/css/mjb-style.css', array('mjb-shared'), MJB_VERSION);
 
         if ($this->should_enqueue_recaptcha()) {
             wp_enqueue_script('google-recaptcha', 'https://www.google.com/recaptcha/api.js', array(), null, true);